experiments

+++ src/main.cpp
@@ -79,7 +79,7 @@ void run_experiment(input &par) {
 
         solution out = alg.run_experiment(par.algorithm);
 
-        if (out.total_energy > par.drone_battery) {
+        if (static_cast<int>(out.total_energy) > par.drone_battery) {
             cerr << "Error: Used energy is=" << out.total_energy << " while total budget is=" << par.drone_battery << endl;
 	        exit(-1);
         }
